refactor: fix the layering and naming the split introduced
Three problems this PR created, all found by review.

**`src/lib` no longer imports from `src/components`.** Moving
`parseBeadsArtifact` into `src/lib/beadsDocument.ts` brought a runtime import of
`tryParseStructuredContent` from `components/workspace/phaseArtifactTypes.ts`
with it, inverting the dependency direction: anything wanting only
`BEADS_APPROVAL_FOCUS_EVENT` was put on a static path through the 2,000-line
workspace parser module. No cycle today, but it becomes one the moment
`phaseArtifactTypes` needs bead parsing. The twelve-line JSON-then-YAML reader
now lives in `src/lib/structuredContent.ts` and both sides import it;
`phaseArtifactTypes` re-exports it so its own parsers are unchanged.

**Two different contracts were sharing the name `ParsedBead`.** The one this PR
put in `lib` is the raw wire shape with every field optional and both spellings
of each; `BeadsApprovalEditor` already exported a `ParsedBead` that is the
normalized editor shape with those fields mandatory. Renamed the new one to
`RawBead`, so a field added to either cannot silently look interchangeable with
the other. Which of the two the app should converge on is a real question, and
not this PR's to answer — naming them apart is what stops the drift meanwhile.

**Two public duration formatters that disagree.** `formatDuration` was
module-private in `helpers.ts` until the split exported it so `openCodeStream`
could reach it, landing it in the barrel beside `formatDurationMs` — which
renders the same input differently (`1m 30s` versus `1.5m`). Renamed to
`formatStreamEventDuration`, with a comment on each saying the other exists.

Also finishes the importer repointing: `ManualQAView` was taking
`CollapsibleSection` from the new module but still pulling
`ArtifactProcessingNotice` from the dispatcher, so it kept the whole viewer in
its graph and got none of the benefit. The three files still importing from the
dispatcher need the dispatcher itself.

Impact: no behaviour change. 1,897 tests across the client, workflow and
structured-output suites pass.
