+++ src/lib/beadsDocument.ts
@@ -1,19 +1,25 @@
 import type { CommandSpec } from '@shared/commandSpec'
 import type { ManualQaBeadOrigin } from '@/hooks/useTickets'
-import { tryParseStructuredContent } from '@/components/workspace/phaseArtifactTypes'
+import { tryParseStructuredContent } from './structuredContent'
 
 export const BEADS_APPROVAL_FOCUS_EVENT = 'beads-approval-focus'
 
 /**
- * A bead as it arrives from the server, with both the camelCase and snake_case
- * spellings the artifact has carried over time.
+ * A bead exactly as it arrives from the server, with both the camelCase and
+ * snake_case spellings the artifact has carried over time and every field
+ * optional.
+ *
+ * Named `RawBead`, not `ParsedBead`: `BeadsApprovalEditor` exports a different
+ * `ParsedBead` — the normalized shape its editor requires, with those fields
+ * mandatory. Two contracts under one name would let a field added to either
+ * side look interchangeable with the other.
  *
  * The artifact viewer declared this and its parser locally, and
  * `BeadsApprovalNavigator` reached for the same records through a bare
  * `bead as Record<string, unknown>` cast — so the outline and the detail view
  * could disagree about what a bead is.
  */
-export interface ParsedBead {
+export interface RawBead {
   [key: string]: unknown
   id?: string
   title?: string
@@ -64,17 +70,17 @@ export interface ParsedBead {
   qa_origin?: ManualQaBeadOrigin | null
 }
 
-export function parseBeadsArtifact(content: string): ParsedBead[] | null {
+export function parseBeadsArtifact(content: string): RawBead[] | null {
   const parsed = tryParseStructuredContent(content)
   if (Array.isArray(parsed)) {
-    return parsed as ParsedBead[]
+    return parsed as RawBead[]
   }
-  if (parsed && typeof parsed === 'object' && !Array.isArray(parsed) && Array.isArray((parsed as { beads?: ParsedBead[] }).beads)) {
-    return (parsed as { beads: ParsedBead[] }).beads
+  if (parsed && typeof parsed === 'object' && !Array.isArray(parsed) && Array.isArray((parsed as { beads?: RawBead[] }).beads)) {
+    return (parsed as { beads: RawBead[] }).beads
   }
   if (content.trim().startsWith('{')) {
     try {
-      return content.trim().split('\n').map((line) => JSON.parse(line) as ParsedBead)
+      return content.trim().split('\n').map((line) => JSON.parse(line) as RawBead)
     } catch {
       return null
     }
